- Fixed a bunch of possible memory leaks and updated the setting of the defaults for the new ivar names in the book Data

+++ Plugins/Source/TBControlDoc.m
@@ -146,7 +146,7 @@
 - (NSString *)filenameFromID:(NSString *)anIdString
 {
 	NSAssert(anIdString != nil, @"anIdString is nil");
-	int markerPos = [anIdString rangeOfString:@"#"].location;
+	NSInteger markerPos = [anIdString rangeOfString:@"#"].location;
 	return (markerPos > 0) ? [anIdString substringToIndex:markerPos] : anIdString;
 	
 }
